+++ src/main/java/com/seong/shoutlink/domain/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.global.exception;
+package com.seong.shoutlink.domain.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
