+++ src/main/java/com/seong/shoutlink/domain/exception/ShoutLinkException.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.global.exception;
+package com.seong.shoutlink.domain.exception;
 
 import lombok.Getter;
