+++ src/main/java/com/seong/shoutlink/global/auth/authentication/JwtAuthenticationInterceptor.java
@@ -1,7 +1,7 @@
 package com.seong.shoutlink.global.auth.authentication;
 
-import com.seong.shoutlink.global.exception.ErrorCode;
-import com.seong.shoutlink.global.exception.ShoutLinkException;
+import com.seong.shoutlink.domain.exception.ErrorCode;
+import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Objects;
