+++ src/main/java/com/seong/shoutlink/global/auth/resolver/LoginUserArgumentResolver.java
@@ -3,8 +3,8 @@ package com.seong.shoutlink.global.auth.resolver;
 import com.seong.shoutlink.domain.auth.LoginUser;
 import com.seong.shoutlink.global.auth.authentication.Authentication;
 import com.seong.shoutlink.global.auth.authentication.AuthenticationContext;
-import com.seong.shoutlink.global.exception.ErrorCode;
-import com.seong.shoutlink.global.exception.ShoutLinkException;
+import com.seong.shoutlink.domain.exception.ErrorCode;
+import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
