+++ src/test/java/com/seong/shoutlink/domain/auth/service/AuthServiceTest.java
@@ -13,8 +13,8 @@ import com.seong.shoutlink.domain.member.Member;
 import com.seong.shoutlink.domain.member.MemberRole;
 import com.seong.shoutlink.domain.member.repository.StubMemberRepository;
 import com.seong.shoutlink.global.auth.jwt.JJwtProvider;
-import com.seong.shoutlink.global.exception.ErrorCode;
-import com.seong.shoutlink.global.exception.ShoutLinkException;
+import com.seong.shoutlink.domain.exception.ErrorCode;
+import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
