+++ src/test/java/com/seong/shoutlink/global/auth/authentication/JwtAuthenticationInterceptorTest.java
@@ -7,8 +7,8 @@ import com.seong.shoutlink.domain.auth.JwtProvider;
 import com.seong.shoutlink.domain.auth.service.response.TokenResponse;
 import com.seong.shoutlink.domain.member.MemberRole;
 import com.seong.shoutlink.fixture.AuthFixture;
-import com.seong.shoutlink.global.exception.ErrorCode;
-import com.seong.shoutlink.global.exception.ShoutLinkException;
+import com.seong.shoutlink.domain.exception.ErrorCode;
+import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
