+++ src/test/java/com/seong/shoutlink/global/auth/resolver/LoginUserArgumentResolverTest.java
@@ -9,8 +9,8 @@ import com.seong.shoutlink.base.BaseControllerTest;
 import com.seong.shoutlink.domain.auth.service.response.TokenResponse;
 import com.seong.shoutlink.domain.member.MemberRole;
 import com.seong.shoutlink.fixture.AuthFixture;
-import com.seong.shoutlink.global.exception.ErrorCode;
-import com.seong.shoutlink.global.exception.ShoutLinkException;
+import com.seong.shoutlink.domain.exception.ErrorCode;
+import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
